workaround libgcrypt pcks padding problems
git-svn-id: https://gnunet.org/svn/GNUnet@407 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/util/hostkey_openssl.c
@@ -373,7 +373,7 @@ int encryptPrivateKey(const void * block,
 int decryptPrivateKey(const struct PrivateKey * hostkey, 
 		      const RSAEncryptedData * block,
 		      void * result,
-		      unsigned int max) {
+		      unsigned short max) {
   RSAEncryptedData tmp; /* this is as big as the result can possibly get */
   int size;
